Change CMAKE_SOURCE_DIR to PROJECT_SOURCE_DIR
This change enables adding Thor as a subproject of another CMake project. This is extremely useful when adding Thor as a git submodule of another projects. It allows the automatic compilation of Thor as part of the recursive project structure without messing it up for people who want to only build it the normal way.

This especially makes sense for people who want to keep tracking Thor from git and directly use it in their projects as well without installing Thor into the system.