[Input] Clarified semantics of ActionContext::event for logical operators
An action consisting of multiple actions combined with logical operators (in particular &&) no longer leads to multiple callback invocations with different ActionContexts. If there is an event action and possibly one or multiple realtime actions, the ActionContext will be filled with the event. The documentation of the corresponding parts has also been improved.